Isolate helper functions.
